Introduced support of SNAPSHOT version of Scala-Native
Scala-Native team made a huge job and right now it is possibly to build
scala-2.12 application on scala-native.

This isn't released feature, but `-SNAPSHOT` version is supported it.

Anyway, in near future (I hope) scalatest needs to add a new version of SN
because current master isn't binary compatibly with `0.4.0-M2`.

This commit:
 - introduced support of specified Scala-Native version via environment variables;
 - cleaned up `NativeBuild.scala` to support both versions